Module podman_ssh_auth_lib::secrets_always_local_mod

Modules

decrypt_mod 🔒

docker_hub_mod 🔒

Push to docker-hub needs the docker hub secret_token. This is a secret important just like a password. I don’t want to pass this secret to an “obscure” library crate that is difficult to review. This secret will stay here in this codebase that every developer can easily inspect. Instead of the secret_token, I will pass the struct DockerHubClient with the trait SendToDockerHub. This way, the secret_token will be encapsulated.

encrypt_mod 🔒

secrecy_mod 🔒

The crate secrecy is probably great. But I want to encrypt the content, so I will make a wrapper. The secrets must always be moved to secrecy types as soon as possible.

ssh_mod 🔒

Structs

EncryptedString

A simple wrapper new-type around String just to show intent that it is already encrypted

Constants

BLUE

ANSI color

GREEN

ANSI color

RED

Secrets like GitHub API secret_token, crates.io secret token, docker hub secret_token, SSH private key passphrase and similar must never go out of this crate. Never pass any secret to an external crate library as much as possible. The user has the source code under his fingers in this crate. So he knows nobody will mess with this code once he inspected and reviewed it. All the modules are in one file to avoid clutter in the automation_tasks_rs folder. The simple program flow of functions that need secrets is butchered to avoid secrets leaving this crate. Now it looks like a mess, but the goal is achieved. The secrets never leave this crate. ANSI color

RESET

ANSI color

YELLOW

ANSI color

Traits

ExposeSecret

Expose a reference to an inner secret