Module podman_ssh_auth_lib::secrets_always_local_mod::docker_hub_mod

Push to docker-hub needs the docker hub token. This is a secret important just like a password. I don’t want to pass this secret to an “obscure” library crate that is difficult to review. This secret will stay here in this codebase that every developer can easily inspect. Instead of the token, I will pass the struct DockerHubClient with the trait SendToDockerHub. This way, the secret token will be encapsulated.

Structs

DockerHubClient

Struct DockerHubClient contains only private fields This fields are accessible only to methods in implementation of traits.